+++ CHANGES.md
@@ -1,5 +1,8 @@
 v0.11.11 - TBD
 =====================
+* Silence an unused variable warning.
+* Remove references to ccall() from the Empscripten build.
+* Improve Android detection.
 
 
 v0.11.10 - 2022-10-20